requirements.txt
langchain>=0.1.0
langchain_community>=0.0.13
fastapi>=0.104.0
uvicorn>=0.24.0
langserve>=0.0.32
streamlit>=1.28.0
python-dotenv>=1.0.0
ollama>=0.1.6
requests>=2.31.0
pytest>=7.4.3
pytest-mock>=3.12.0
huggingface-hub>=0.19.0
transformers>=4.35.0 